+++ projects/app/src/pages/api/core/dataset/data/delete.ts
@@ -3,7 +3,7 @@ import { jsonRes } from '@fastgpt/service/common/response';
 import { withNextCors } from '@fastgpt/service/common/middle/cors';
 import { connectToDatabase } from '@/service/mongo';
 import { authDatasetData } from '@/service/support/permission/auth/dataset';
-import { deleteDataByDataId } from '@/service/core/dataset/data/controller';
+import { delDatasetDataByDataId } from '@fastgpt/service/core/dataset/data/controller';
 
 export default withNextCors(async function handler(req: NextApiRequest, res: NextApiResponse<any>) {
   try {
@@ -19,7 +19,7 @@ export default withNextCors(async function handler(req: NextApiRequest, res: Nex
     // 凭证校验
     await authDatasetData({ req, authToken: true, dataId, per: 'w' });
 
-    await deleteDataByDataId(dataId);
+    await delDatasetDataByDataId(dataId);
 
     jsonRes(res, {
       data: 'success'
