+++ projects/app/src/web/support/wallet/pay/api.ts
@@ -1,6 +1,6 @@
 import { GET } from '@/web/common/api/request';
 import type { PaySchema } from '@fastgpt/global/support/wallet/pay/type.d';
-import { delay } from '@/utils/tools';
+import { delay } from '@fastgpt/global/common/system/utils';
 
 export const getPayOrders = () => GET<PaySchema[]>(`/plusApi/support/wallet/pay/getPayOrders`);
